requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.0
mkl-service==2.3.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1643958805922/work
oauthlib==3.2.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.5
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.63.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0